% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Simple centering helper}
\usage{
center(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric variable.}

\item{na.rm}{Remove missing values (default TRUE)?}
}
\value{
Numeric variable centered around its mean.
}
\description{
Simple centering helper
}
\examples{
center(c(1, 2, 3, 5, 3, 2), na.rm = TRUE)
center(c(1, NA, NA, 5, 3, 2), na.rm = TRUE)
}
