% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sem_fit.R
\name{get_sem_fit}
\alias{get_sem_fit}
\title{Obtain lavaan model fit indices}
\usage{
get_sem_fit(model, round = 6)
}
\arguments{
\item{model}{Fitted lavaan model.}

\item{round}{Rounding for the numeric output.}
}
\value{
Fit indices and model statistics.
}
\description{
Obtain lavaan model fit indices
}
\examples{
set.seed(21417)
x <- rnorm(1000)

d <- data.frame(
  x1 = sqrt(.50) * x + sqrt(1 - .50) * rnorm(1000),
  x2 = sqrt(.50) * x + sqrt(1 - .50) * rnorm(1000),
  x3 = sqrt(.50) * x + sqrt(1 - .50) * rnorm(1000),
  x4 = sqrt(.50) * x + sqrt(1 - .50) * rnorm(1000)
)

mod <- "X=~x1+x2+x3+x4"

fit1 <- lavaan::cfa(mod, d)

get_sem_fit(fit1, round = 3)
}
