% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mean_center.R
\name{group_mean_center}
\alias{group_mean_center}
\title{Returns data including group means and group mean centered variables}
\usage{
group_mean_center(data, group.var, vars, grand.init = FALSE)
}
\arguments{
\item{data}{data frame of individual observations across multiple groups}

\item{group.var}{character string name of the grouping variable}

\item{vars}{vector of variable names to be centered}

\item{grand.init}{is initial grand mean centering to each variable applied (default FALSE)}
}
\value{
data frame with group means and variables centered around group means
}
\description{
Returns data including group means and group mean centered variables
}
\examples{
set.seed(234)
n1=3 #groups
n2=3 #observations per groups
dat<-data.frame(
group=rep(c(LETTERS[1:n1]),each=n2),
x1=sample(1:5,9,replace=TRUE),
x2=sample(1:5,9,replace=TRUE),
x3=sample(1:5,9,replace=TRUE))

group_mean_center(data=dat,group.var = "group",
vars = c("x1","x2","x3"))
}
