% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_standardize.R
\name{na_standardize}
\alias{na_standardize}
\title{Standardizing helper}
\usage{
na_standardize(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric variable.}

\item{na.rm}{Logical. Should missing values be removed. Default TRUE.}
}
\value{
Standardized numeric variable.
}
\description{
Standardizing helper
}
\examples{
na_standardize(c(1,4,2,5,7))
}
