% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{notin}
\alias{notin}
\title{Value mismatching}
\usage{
notin(x, y)
}
\arguments{
\item{x}{vector: the values to be matched}

\item{y}{vector: the values to be matched against}
}
\value{
a vector of positions of mismatches
}
\description{
Value mismatching
}
\examples{
notin(c("a"),c("a","b"))
notin(c("a"),c("c","b"))
}
