% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tost_ml.R
\name{tost_ml}
\alias{tost_ml}
\title{Two One-Sided Tests for Fixed Effects in Multilevel Models}
\usage{
tost_ml(model, fe, bound_l, bound_u)
}
\arguments{
\item{model}{Multilevel model object fitted with lmerTest::lmer.}

\item{fe}{Character vector. Name of the fixed effect variable.}

\item{bound_l}{Lower bound based on SESOI (smallest effect size of interest).}

\item{bound_u}{Upper bound based on SESOI (smallest effect size of interest).}
}
\value{
Vector of test statistics against lower and upper bound.
}
\description{
Two One-Sided Tests for Fixed Effects in Multilevel Models
}
\examples{
set.seed(234)
n1 <- 3 # groups
n2 <- 3 # observations per groups

dat <- data.frame(
  group = rep(c(LETTERS[1:n1]), each = n2),
  x1 = sample(1:5, n1 * n2, replace = TRUE),
  x2 = sample(1:5, n1 * n2, replace = TRUE),
  x3 = sample(1:5, n1 * n2, replace = TRUE)
)

tost_ml(
  model = lmerTest::lmer(x1 ~ x2 + x3 + (1 | group),
    data = dat, REML = FALSE
  ),
  fe = "x2", bound_l = -0.30, bound_u = 0.30
)
}
