% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Simple centering helper}
\usage{
center(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric variable}

\item{na.rm}{remove missing values? (default TRUE)}
}
\value{
numeric variable centered around its mean
}
\description{
Simple centering helper
}
\examples{
center(c(1,2,3,5,3,2),na.rm=TRUE)
}
